+++ distribution/libs/distribution/include/distribution/distribution.h
@@ -5,7 +5,7 @@
 class Distribution {
   public:
     static std::vector<double> GenerateRandomDistribution(int length);
-    static std::vector<double> GenerateUniformDistribution(int a, int b,
+    static std::vector<double> GenerateUniformDistribution(double a, double b,
                                                            int length);
     static std::vector<double>
     GenerateGaussDistribution(double mean, double std, int n, int length);
@@ -13,10 +13,10 @@ class Distribution {
                                                                int length);
     static std::vector<double> GenerateGammaDistribution(int eta, double param,
                                                          int length);
-    static std::vector<double> GenerateMinTriangleDistribution(int a, int b,
+    static std::vector<double> GenerateMinTriangleDistribution(double a, double b,
                                                                int length);
-    static std::vector<double> GenerateMaxTriangleDistribution(int a, int b,
+    static std::vector<double> GenerateMaxTriangleDistribution(double a, double b,
                                                                int length);
-    static std::vector<double> GenerateSimpsonDistribution(int a, int b,
+    static std::vector<double> GenerateSimpsonDistribution(double a, double b,
                                                            int length);
 };
